refactor: change SelectedDevicesRepository from class to object
