+++ src/data/selected-devices/SelectedDevicesRepository.ts
@@ -2,16 +2,11 @@ import { JSONtoMap, maptoJSON } from "../../utils/helpers/jsonHelpers";
 import { DeviceName } from "../constants/DeviceEmissions";
 import { IStorageRepository } from "../storage/IStorageRepository";
 import { StorageKeys } from "../storage/StorageKeys";
-import {
-    ISelectedDevicesRepository,
-    isSelectedDevicesMap,
-} from "./ISelectedDevicesRepository";
+import { isSelectedDevicesMap } from "./DeviceNames";
 
-export class SelectedDevicesRepository implements ISelectedDevicesRepository {
-    remoteDataSource: IStorageRepository = IStorageRepository.instance;
-
-    async getSelectedDevices(): Promise<Map<DeviceName, number>> {
-        const data = await this.remoteDataSource.get<string>(
+export const SelectedDevicesRepository = {
+    getSelectedDevices: async function (): Promise<Map<DeviceName, number>> {
+        const data = await IStorageRepository.instance.get<string>(
             StorageKeys.selectedDevices,
             maptoJSON(new Map<DeviceName, number>([]))
         );
@@ -21,35 +16,35 @@ export class SelectedDevicesRepository implements ISelectedDevicesRepository {
             throw Error(`data ${data} is not a valid map`);
         }
         return map;
-    }
+    },
 
-    async addSelectedDevice(device: DeviceName): Promise<void> {
+    addSelectedDevice: async function (device: DeviceName): Promise<void> {
         const newMap = await this.getSelectedDevices();
 
         if (!newMap.has(device)) {
             newMap.set(device, 0);
         }
 
-        await this.remoteDataSource.set(
+        await IStorageRepository.instance.set(
             StorageKeys.selectedDevices,
             maptoJSON(newMap)
         );
-    }
+    },
 
-    async removeSelectedDevice(device: DeviceName): Promise<void> {
+    removeSelectedDevice: async function (device: DeviceName): Promise<void> {
         const newMap = await this.getSelectedDevices();
 
         if (newMap.has(device)) {
             newMap.delete(device);
         }
 
-        await this.remoteDataSource.set(
+        await IStorageRepository.instance.set(
             StorageKeys.selectedDevices,
             maptoJSON(newMap)
         );
-    }
+    },
 
-    async setSelectedDevicePercentage(
+    setSelectedDevicePercentage: async function (
         device: DeviceName,
         percentage: number
     ): Promise<void> {
@@ -63,9 +58,9 @@ export class SelectedDevicesRepository implements ISelectedDevicesRepository {
             );
         }
 
-        await this.remoteDataSource.set(
+        await IStorageRepository.instance.set(
             StorageKeys.selectedDevices,
             maptoJSON(newMap)
         );
-    }
-}
+    },
+};
